// types.ts
export enum ConverterType {
  CaseConverter = 'caseConverter',
  ImageCompressor = 'imageCompressor',
  PdfToText = 'pdfToText',
  LoremIpsumGenerator = 'loremIpsumGenerator',
}

export enum CaseType {
  Uppercase = 'uppercase',
  Lowercase = 'lowercase',
  TitleCase = 'titlecase',
  SentenceCase = 'sentencecase',
}